// src/redux/video-metadata/video-metadata.enum.ts
export enum VIDEO_METADATA_TYPES {
  SET_VIDEO_ID = 'SET_VIDEO_ID',
  SET_VIDEO_THUMBNAIL_SRC = 'SET_VIDEO_THUMBNAIL_SRC',
  SET_VIDEO_TITLE = 'SET_VIDEO_TITLE',
  SET_VIDEO_DESCRIPTION = 'SET_VIDEO_DESCRIPTION',
  SET_VIDEO_LENGTH_SECONDS = 'SET_VIDEO_LENGTH_SECONDS',
  SET_VIDEO_POST_DATE = 'SET_VIDEO_POST_DATE',
  SET_VIDEO_ALREADY_WATCHED = 'SET_VIDEO_ALREADY_WATCHED',
  SET_VIDEO_HASHTAGS = 'SET_VIDEO_HASHTAGS',
  // Set entire state
  SET_VIDEO_METADATA_STATE = 'SET_VIDEO_METADATA_STATE',
  RESET_VIDEO_METADATA_STATE = 'RESET_VIDEO_METADATA_STATE',
}